pyproject.toml
[tool.poetry]
name = "vision-stream"
version = "0.1.0"
description = ""
authors = ["dnl-strnl <129630837+dnl-strnl@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "vision_stream", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
hydra-core = "^1.3.2"
ultralytics = "^8.3.56"
torch-model-archiver = "^0.12.0"
torchserve = "0.8.1"
nvgpu = "^0.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
